requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Pillow==10.1.0
twilio==8.10.0
python-dotenv==1.0.0
werkzeug==3.0.1